feat: add PdfReader.tsx
