+++ next.config.js
@@ -1,8 +1,15 @@
 /** @type {import('next').NextConfig} */
 module.exports = {
     reactStrictMode: false,
-    runtime: 'edge',
+    runtime: "edge",
     experimental: {
         appDir: true
+    },
+    webpack: (config) => {
+        // Avoid: Can't resolve 'canvas'
+        config.externals.push({
+            canvas: "canvas"
+        });
+        return config;
     }
-}
+};
