+++ package.json
@@ -23,25 +23,27 @@
     "@piotr-cz/swr-idb-cache": "^1.0.0",
     "@radix-ui/colors": "^0.1.8",
     "@radix-ui/react-toast": "^1.1.2",
+    "@react-pdf-viewer/core": "^3.10.0",
+    "@react-pdf-viewer/default-layout": "^3.10.0",
     "dropbox": "^10.34.0",
     "encoding": "^0.1.13",
     "exponential-backoff-generator": "^2.0.1",
     "husky": "^8.0.3",
     "msw": "^0.49.2",
-    "next": "^13.1.1",
+    "next": "^13.1.4",
+    "pdfjs-dist": "3.1.81",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-reader": "^0.21.2",
     "react-use": "^17.4.0",
-    "swr": "^2.0.0"
+    "swr": "^2.0.1"
   },
   "devDependencies": {
     "@types/node": "^18.11.18",
-    "@types/react": "^18.0.26",
+    "@types/react": "^18.0.27",
     "eslint": "^8.31.0",
-    "eslint-config-next": "^13.1.1",
+    "eslint-config-next": "^13.1.4",
     "lint-staged": "^13.1.0",
-    "prettier": "^2.8.1",
+    "prettier": "^2.8.3",
     "ts-node-test": "^0.3.0",
     "typescript": "^4.9.4"
   },
